% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means.R
\name{plot_means}
\alias{plot_means}
\title{Plot Results of Impact Analysis}
\usage{
plot_means(
  data,
  font = "Lato",
  title = NULL,
  subtitle = NULL,
  blabel_size = 20,
  label_adjust = 0,
  xlabs,
  colors = c("#e5daa5", "#90310b", "#90315a"),
  groupvar,
  outcomevar
)
}
\arguments{
\item{subtitle}{The subtitle for the plot.}

\item{dataset}{Data frame containing the dataset.}

\item{outcomes}{Character vector of outcome variable names.}

\item{treatment_var}{The name of the treatment variable.}

\item{fixed_effect_var}{The name of the fixed effect variable (optional).}

\item{cluster_var}{The name of the cluster variable (optional).}

\item{level}{The level for confidence intervals.}

\item{color}{bar color}

\item{treatment_labs}{Character vector of labels for treatment groups.}

\item{legend_title}{The title for the legend.}

\item{outcome_labs}{Character vector of labels for outcome variables.}
}
\value{
A ggplot object representing the impact analysis results.
}
\description{
This function plots the results of an impact analysis using the ImpactViz package.
}
\examples{
plot_results(dataset = your_dataset,
             outcomes = c("outcome1", "outcome2"),
             treatment_var = "your_treatment_var",
             colors = c("grey", "blue"),
             treatment_labs = c("Control", "Treatment"),
             subtitle = "Your Subtitle",
             legend_title = "Treatment Group")
}
